add email registered form
