+++ lib/pages/emailRegistered/createUserWithEmailAndPasswordWidget.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:readr_app/blocs/emailRegistered/bloc.dart';
-import 'package:readr_app/blocs/emailRegistered/events.dart';
 import 'package:readr_app/blocs/emailRegistered/states.dart';
-import 'package:readr_app/helpers/dataConstants.dart';
 import 'package:readr_app/pages/emailRegistered/emailRegisteredErrorWidget.dart';
+import 'package:readr_app/pages/emailRegistered/emailRegisteredForm.dart';
 import 'package:readr_app/pages/emailRegistered/emailValidatorWidget.dart';
 import 'package:readr_app/pages/emailRegistered/passwordValidatorWidget.dart';
 import 'package:readr_app/widgets/memberLoginPolicy.dart';
@@ -23,52 +21,6 @@ class CreateUserWithEmailAndPasswordWidget extends StatefulWidget {
 class _CreateUserWithEmailAndPasswordWidgetState extends State<CreateUserWithEmailAndPasswordWidget> {
   final _emailEditingController = TextEditingController();
   final _passwordEditingController = TextEditingController();
-  bool _isHidden = true;
-  bool _emailIsValid = false;
-  bool _passwordIsValid = false;
-
-  @override
-  void initState() {
-    _emailEditingController.text = widget.email;
-    _emailIsValid = _isEmailValid();
-    _emailEditingController.addListener(
-      () {
-        setState(() {
-          _emailIsValid = _isEmailValid();
-        });
-      }
-    );
-    _passwordEditingController.addListener(
-      () {
-        setState(() {
-          _passwordIsValid = _isPasswordValid();
-        });
-      }
-    );
-    super.initState();
-  }
-
-  bool _isEmailValid() {
-    Pattern pattern = r"^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$";
-    RegExp regex = RegExp(pattern);
-    return _emailEditingController.text != null && regex.hasMatch(_emailEditingController.text);
-  }
-
-  bool _isPasswordValid() {
-    return _passwordEditingController.text != null && _passwordEditingController.text.length >= 6;
-  }
-
-  _createUserWithEmailAndPassword(String email, String password) async {
-    context.read<EmailRegisteredBloc>().add(
-      CreateUserWithEmailAndPassword(email: email, password: password)
-    );
-  }
-
-  void _togglePasswordView() {
-    setState(() {
-      _isHidden = !_isHidden;
-    });
-  }
 
   void _delayNavigatorPop() async{
     await Future.delayed(Duration());
@@ -97,171 +49,13 @@ class _CreateUserWithEmailAndPasswordWidgetState extends State<CreateUserWithEma
         }
 
         // state is Init, Loading, or EmailAlreadyInUse 
-        return _buildEmailRegisteredBlock(state);
+        return EmailRegisteredForm(
+          email: widget.email,
+          state: state,
+          emailEditingController: _emailEditingController,
+          passwordEditingController: _passwordEditingController,
+        );
       }
     );
   }
-
-  Widget _buildEmailRegisteredBlock(EmailRegisteredState state) {
-    bool isLoading = state is EmailRegisteredLoading;
-
-    return CustomScrollView(
-      slivers: [
-        SliverToBoxAdapter(child: SizedBox(height: 56)),
-        SliverToBoxAdapter(
-          child: Padding(
-            padding: const EdgeInsets.only(left: 24.0, right: 24.0),
-            child: _emailTextField(),
-          ),
-        ),
-        SliverToBoxAdapter(child: SizedBox(height: 8)),
-        SliverToBoxAdapter(
-          child: Padding(
-            padding: const EdgeInsets.only(left: 24.0, right: 24.0),
-            child: EmailValidatorWidget(editingController: _emailEditingController,),
-          ),
-        ),
-        SliverToBoxAdapter(child: SizedBox(height: 32)),
-        SliverToBoxAdapter(
-          child: Padding(
-            padding: const EdgeInsets.only(left: 24.0, right: 24.0),
-            child: _passwordField(),
-          ),
-        ),
-        SliverToBoxAdapter(child: SizedBox(height: 8)),
-        SliverToBoxAdapter(
-          child: Padding(
-            padding: const EdgeInsets.only(left: 24.0, right: 24.0),
-            child: PasswordValidatorWidget(editingController: _passwordEditingController,),
-          ),
-        ),
-        SliverToBoxAdapter(child: SizedBox(height: 24)),
-        if(state is EmailAlreadyInUse)
-        ...[
-          SliverToBoxAdapter(
-            child: Center(
-              child: Text(
-                '這個 Email 已經註冊過囉',
-                style: TextStyle(
-                  fontSize: 16,
-                  color: Colors.red
-                ),
-              ),
-            ),
-          ),
-          SliverToBoxAdapter(child: SizedBox(height: 12)),
-        ],
-        SliverToBoxAdapter(
-          child: Padding(
-            padding: const EdgeInsets.only(left: 24.0, right: 24.0),
-            child: isLoading
-            ? _emailLoadingButton()
-            : _registerButton(_emailIsValid && _passwordIsValid),
-          ),
-        ),
-        SliverToBoxAdapter(child: SizedBox(height: 24)),
-        SliverFillRemaining(
-          hasScrollBody: false,
-          child: Align(
-            alignment: Alignment.bottomCenter,
-            child: Padding(
-              padding: const EdgeInsets.only(bottom: 44.0),
-              child: Container(height: 50, child: MemberLoginPolicy()),
-            )
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _emailTextField() {
-    return Form(
-      autovalidateMode: AutovalidateMode.onUserInteraction,
-      child: TextFormField(
-        controller: _emailEditingController,
-        style: TextStyle(
-          color: Colors.black,
-          fontSize: 17,
-        ),
-        decoration: InputDecoration(
-          labelText: 'Email',
-          labelStyle: TextStyle(
-            color: Colors.black,
-            fontSize: 17,
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _passwordField() {
-    return Form(
-      child: TextFormField(
-        controller: _passwordEditingController,
-        obscureText: _isHidden,
-        style: TextStyle(
-          color: Colors.black,
-          fontSize: 17,
-        ),
-        decoration: InputDecoration(
-          suffixIcon: IconButton(
-            onPressed: () => _togglePasswordView(),
-            icon: _isHidden 
-            ? Icon(Icons.visibility)
-            : Icon(Icons.visibility_off),
-          ),
-          labelText: '設定密碼',
-          labelStyle: TextStyle(
-            color: Colors.black,
-            fontSize: 17,
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _emailLoadingButton() {
-    return Container(
-      height: 50.0,
-      decoration: BoxDecoration(
-        color: appColor,
-        borderRadius: BorderRadius.circular(5.0),
-      ),
-      child: SpinKitThreeBounce(color: Colors.white, size: 35,),
-    );
-  }
-  
-  Widget _registerButton(bool emailAndPasswordIsValid) {
-    return InkWell(
-      borderRadius: BorderRadius.circular(5.0),
-      child: Container(
-        height: 50.0,
-        decoration: BoxDecoration(
-          color: emailAndPasswordIsValid
-          ? appColor
-          : Color(0xffE3E3E3),
-          borderRadius: BorderRadius.circular(5.0),
-        ),
-        child: Center(
-          child: Text(
-            '註冊會員',
-            style: TextStyle(
-              fontSize: 17,
-              color: emailAndPasswordIsValid
-              ? Colors.white
-              : Colors.grey
-            ),
-          ),
-        ),
-      ),
-      onTap: emailAndPasswordIsValid
-      ? () async{
-          _createUserWithEmailAndPassword(
-            _emailEditingController.text, 
-            _passwordEditingController.text
-          );
-        }
-      : null
-    );
-  }  
 }
